further fix 68k APD support
The APD timer is decremented by the default autovector 5 handler, and thus we need it be installed a majority of the
time. Unfortunately, we /need/ to mask it while reading the keyboard or else we might read in whatever data is left on
the bus by said handler, instead of the actual keyboard state. For now I've inserted a for loop that calls idle() 8
times, which seems to impart sufficient delay into the main loop for everything to work.

Now I need to clean things up /again/.